every page table except root tracks number of populated PAGE_SIZE chunks

+++ memory.h
@@ -80,7 +80,7 @@ struct page {
 
 	union {
 		struct kmem_slab *slab;
-		int refcount;
+		unsigned long refcount;
 		int order;
 	} u;
 };
